add the script for generating flamegraphs
this takes a pid and generates a flamegraph from it with a couple defaults

it sets up a directory and dumps a ton of intermediate-stage files there
so you can easily debug. makes folded scripts, JSON/burns them etc

also does a questionable thing by combining scripts by catting them together.
this is a little vestigal from earlier versions of the code but useful for the
whole picture view of the graph. some cleanup might be needed